fix(ci): unblock cascading clippy on main
Allow `clippy::too_many_arguments` workspace-wide in
`.github/workflows/clippy.yml` with an inline justification: Soroban
contract entry points receive args directly from the transaction ABI,
and grouping them into structs would change the public method ABI
that existing clients bind against.

Mechanically resolve the cascading `-D correctness -D suspicious
-D complexity` failures that blocked CI on main:

- Fix typo in `puzzle_dao#![allow]`:
  `empty_line_after_outer_attribute` -> `empty_line_after_outer_attr`
  (clippy warned the allow was a no-op, which unblocked 5 deny
  errors in the same file).
- `skill_rating::season_id_for`: rewrite
  `if season_len == 0 {0} else { now/season_len }` as
  `now.checked_div(season_len).unwrap_or(0)` to satisfy
  `clippy::manual_checked_division`.
- Drop redundant `as u32` / `as i128` casts on already-typed operands
  across 11 contracts; Soroban `Vec::len()` returns `u32` and token
  amounts are `i128`.
- Drop redundant `.clone()` on `Copy` types in `multisig_treasury`
  and `achievement_collection`.
- `governance_token` binary search: rewrite
  `let mid = low + (high - low + 1) / 2` as
  `let mid = low + (high - low).div_ceil(2)`.

End-to-end CI matrix (clippy + check + test --no-run + doc) was
verified green except for warn-only lint long-tails, which are
out of scope for this fix.

Also bundles the prior rustfmt-formatting sweep on the rest of the
workspace (`cargo fmt --all`) -- these changes are unrelated to the
clippy cascade but were already in the working tree.

+++ contracts/flash_loan/src/lib.rs
@@ -243,8 +243,7 @@ impl FlashLoanContract {
             .get(&DataKey::Pool(token.clone()))
             .unwrap_or_else(|| panic!("Pool not found for token"));
 
-        let max_loan = (pool.available_liquidity as i128 * config.max_loan_ratio as i128
-            / BASIS_POINTS) as i128;
+        let max_loan = (pool.available_liquidity * config.max_loan_ratio as i128) / BASIS_POINTS;
         if amount > max_loan {
             panic!("Amount exceeds maximum loan limit");
         }
